Semaphore stdout in Java implementation
